Added a mockup of the site in PHP for comparison to the HTML-only version
The PHP version incorporates a number of enhancements over the HTML-only version.
It highlights the current page in the navigation and removes that page's link.
It displays one of a rotating group of mottos.
It separates all generic page contant from specific page content, making administration of the site simpler.
It attempts to detect mobile browsers and display the name of the page shorter.
It does all of these things whether the visitor has JavaScript enabled.